+++ modules/helpers/crew.js
@@ -134,7 +134,7 @@ export async function updateRoles(vehicle_actor, crew_member_id, new_crew_roles)
       'actor_id': crew_member.id,
       'actor_name': crew_member.name,
       'role': newRole,
-      'link':  await TextEditor.enrichHTML(crew_member?.link) || null,
+      'link':  await foundry.applications.ux.TextEditor.enrichHTML(crew_member?.link) || null,
     });
   }
 
@@ -282,7 +282,7 @@ export async function selectRoles(vehicle, crew_member_id) {
   const crewMemberRoles = vehicleRoles.filter(role => role.actor_id === crew_member_id);
   const rolesInUse = crewMemberRoles.map(role => role.role);
 
-  const content = await renderTemplate(
+  const content = await foundry.applications.handlebars.renderTemplate(
     "systems/starwarsffg/templates/dialogs/ffg-crew-change.html",
     {
       actor: crew_member,