+++ modules/items/item-sheet-ffg.js
@@ -92,11 +92,11 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
     }
 
     if (data?.data?.description) {
-      data.data.enrichedDescription = await TextEditor.enrichHTML(data.data.description);
+      data.data.enrichedDescription = await foundry.applications.ux.TextEditor.enrichHTML(data.data.description);
     }
 
     if (data?.data?.longDesc !== undefined) {
-      data.data.enrichedLongDesc = await TextEditor.enrichHTML(data.data.longDesc);
+      data.data.enrichedLongDesc = await foundry.applications.ux.TextEditor.enrichHTML(data.data.longDesc);
       data.data.hasLongDesc = true;
     } else {
       data.data.hasLongDesc = false;
@@ -169,7 +169,7 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
           data.isReadOnly = true;
         }
         for (let x = 0; x < 16; x++) {
-          data.data.upgrades[`upgrade${x}`].enrichedDescription = await TextEditor.enrichHTML(data.data.upgrades[`upgrade${x}`].description);
+          data.data.upgrades[`upgrade${x}`].enrichedDescription = await foundry.applications.ux.TextEditor.enrichHTML(data.data.upgrades[`upgrade${x}`].description);
         }
         break;
       case "specialization":
@@ -190,7 +190,7 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
           this.item.flags.starwarsffg.loaded = true;
         }
         for (let x = 0; x < 20; x++) {
-          data.data.talents[`talent${x}`].enrichedDescription = await TextEditor.enrichHTML(data.data.talents[`talent${x}`].description);
+          data.data.talents[`talent${x}`].enrichedDescription = await foundry.applications.ux.TextEditor.enrichHTML(data.data.talents[`talent${x}`].description);
         }
         break;
       case "species":
@@ -287,7 +287,7 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
           data.data.isReadOnly = true;
         }
         for (let x = 0; x < 8; x++) {
-          data.data.upgrades[`upgrade${x}`].enrichedDescription = await TextEditor.enrichHTML(data.data.upgrades[`upgrade${x}`].description);
+          data.data.upgrades[`upgrade${x}`].enrichedDescription = await foundry.applications.ux.TextEditor.enrichHTML(data.data.upgrades[`upgrade${x}`].description);
         }
         break;
       }
@@ -424,6 +424,7 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
   activateListeners(html) {
     super.activateListeners(html);
     html.find(".ffg-purchase").click(async (ev) => {
+      if(this.actor && !this.actor?.verifyEditModeIsNotEnabled()) return;
       await this._handleItemBuy(ev)
     });
 
@@ -448,7 +449,7 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
     // Activate tabs
     let tabs = html.find(".tabs");
     let initial = this._sheetTab;
-    new Tabs(tabs, {
+    new foundry.applications.ux.Tabs(tabs, {
       initial: initial,
       callback: (clicked) => (this._sheetTab = clicked.data("tab")),
     });
@@ -597,7 +598,7 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
       html.find(".talent-actions .fa-cog").on("click", ModifierHelpers.popoutModiferWindow.bind(this));
       html.find(".talent-modifiers .fa-cog").on("click", this._onClickUpgradeEdit.bind(this));
       try {
-        const dragDrop = new DragDrop({
+        const dragDrop = new foundry.applications.ux.DragDrop({
           dragSelector: ".item",
           dropSelector: ".specialization-talent",
           permissions: { dragstart: this._canDragStart.bind(this), drop: this._canDragDrop.bind(this) },
@@ -610,7 +611,7 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
       }
     } else if (this.object.type === "career") {
       try {
-        const dragDrop = new DragDrop({
+        const dragDrop = new foundry.applications.ux.DragDrop({
           dragSelector: ".item",
           dropSelector: ".tab.career",
           permissions: { dragstart: this._canDragStart.bind(this), drop: this._canDragDrop.bind(this) },
@@ -656,7 +657,7 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
       });
     } else if (this.object.type === "species") {
       try {
-        const dragDrop = new DragDrop({
+        const dragDrop = new foundry.applications.ux.DragDrop({
           dragSelector: ".item",
           dropSelector: ".tab.talents",
           permissions: { dragstart: this._canDragStart.bind(this), drop: this._canDragDrop.bind(this) },
@@ -715,7 +716,7 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
     });
 
     if (["weapon", "armour", "itemattachment", "shipweapon"].includes(this.object.type)) {
-      const itemToItemAssociation = new DragDrop({
+      const itemToItemAssociation = new foundry.applications.ux.DragDrop({
         dragSelector: ".item",
         dropSelector: null,
         permissions: { dragstart: true, drop: true },
@@ -1313,6 +1314,8 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
 
   /** @override */
   _updateObject(event, formData) {
+    if(this.actor && !this.actor?.verifyEditModeIsNotEnabled()) return;
+
     const itemUpdate = ItemHelpers.itemUpdate.bind(this);
     itemUpdate(event, formData);
   }
@@ -1324,6 +1327,8 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
    * @private
    */
   async _onClickUpgradeEdit(event) {
+    if(this.actor && !this.actor?.verifyEditModeIsNotEnabled()) return;
+
     // pull the item which the edit is on
     const li = $(event.currentTarget);
     const clickedId = li.closest('.talent-block').attr('id');
@@ -1380,6 +1385,8 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
    */
   async _onClickTalentControl(event) {
     event.preventDefault();
+    if(this.actor && !this.actor?.verifyEditModeIsNotEnabled()) return;
+
     const a = event.currentTarget;
     const action = a.dataset.action;
     const key = a.dataset.key;
@@ -1478,7 +1485,7 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
     }
 
     if (action === "img") {
-      const fp = new FilePicker({
+      const fp = new foundry.applications.apps.FilePicker({
         type: "image",
         callback: async (path) => {
           await this.object.update({img: path});
@@ -1944,7 +1951,7 @@ export class ItemSheetFFG extends foundry.appv1.sheets.ItemSheet {
       let details = li.children(".item-details");
       details.slideUp(200, () => details.remove());
     } else {
-      let div = $(`<div class="item-details">${await TextEditor.enrichHTML(desc)}</div>`);
+      let div = $(`<div class="item-details">${await foundry.applications.ux.TextEditor.enrichHTML(desc)}</div>`);
       li.append(div.hide());
       div.slideDown(200);
     }